requirements.txt
xlrd
pandas
pinecone
python-dotenv
sentence-transformers
openai 
streamlit
fpdf